requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
ollama
pytest
pytest-asyncio
pytest-cov
python-dotenv
aiofiles
email-validator
structlog
prometheus-client
psutil
